added some validation

+++ src/automation/3.1_install_kubernetes.sh
@@ -16,66 +16,100 @@ print_instruction " |    \_ |_____| |_____] |______ |    \_ |  \_| |______    | 
 
 . _array_setup.sh
 
-# Run this code on the master
-print_instruction "\nRemoving the swap file on $ip_addr_me\n"
-sudo dphys-swapfile swapoff
-sudo dphys-swapfile uninstall
-sudo update-rc.d dphys-swapfile remove
-sudo apt-get -y purge dphys-swapfile
+doMaster=0
 
-print_instruction "\nUpdate and Upgrade"
-sudo apt update
-sudo apt -y upgrade
+while true; do
+    printf "\n\nDo you want to install Kubernetes on the master node?"
+    read -p "(y/n)?" yn
+    case $yn in
+        [Yy]* )
+            break
+            ;;
+        [Nn]* )
+            doMaster=1
+            break
+            ;;
+        * ) echo "Please answer (y)es or (n)o.";;
+    esac
+done
 
-print_instruction "\nAdding link to Kubernetes repository and adding the APT key\n"
-sudo curl -s https://packages.cloud.google.com/apt/doc/apt-key.gpg | sudo apt-key add -
+if [ doMaster ]
+then
 
-print_instruction "\nModify iptables"
-sudo update-alternatives --set iptables /usr/sbin/iptables-legacy
-sudo update-alternatives --set ip6tables /usr/sbin/ip6tables-legacy
-sudo update-alternatives --set ebtables /usr/sbin/ebtables-legacy
-
-print_instruction "\nCreating support file for k8s: kubernetes.list"
-# Create a support file that will be copied to the nodes
-echo "deb http://apt.kubernetes.io/ kubernetes-xenial main" | tee kubernetes.list
-
-print_instruction "\nAdd Kubernetes repository to the RPi package lists"
-sudo rm -f /etc/apt/sources.list.d/kubernetes.list
-sudo cp -f kubernetes.list /etc/apt/sources.list.d/kubernetes.list
-sudo echo 'Acquire::https::packages.cloud.google.com::Verify-Peer "false";' > apt.conf
-sudo rm -f /etc/apt/apt.conf
-sudo cp -f apt.conf /etc/apt/
-
-print_instruction "\nUpdate & install kubelet kubeadm kubectl"
-x=1
-while [ $x -le 5 ]
-do
-    print_instruction "Updating..."
-    count=sudo apt update | grep -c "404  Not Found"
-
-    if (( count >= 0 ))
+    if [ $(swapon --show | grep -c "NAME") -gt 0 ]
     then
-        x=10
+        # Run this code on the master
+        print_instruction "\nRemoving the swap file on $ip_addr_me\n"
+        sudo dphys-swapfile swapoff
+        sudo dphys-swapfile uninstall
+        sudo update-rc.d dphys-swapfile remove
+        sudo apt-get -y purge dphys-swapfile
     fi
 
-    x=$(( $x + 1 ))
+    if [ $(swapon --show | grep -c "NAME") -gt 0 ]
+    then
+        print_instruction "\nDiabling the swap file did not work...stopping."
+        exit 2
+    fi
 
-done
-sudo apt install -y kubelet kubeadm kubectl
+    print_instruction "\nUpdate and Upgrade"
+    sudo apt update
+    sudo apt -y upgrade
 
-print_instruction "\nkubeadm init...\n"
-sudo kubeadm init --ignore-preflight-errors=all --pod-network-cidr 10.244.0.0/16 --apiserver-advertise-address=$ip_addr_me
+    print_instruction "\nAdding link to Kubernetes repository and adding the APT key\n"
+    sudo curl -s https://packages.cloud.google.com/apt/doc/apt-key.gpg | sudo apt-key add -
 
-print_instruction "Configuring Kubernetes with local user"
-echo $piid > piid.txt
-sudo -u $piid sh ./_kube_config.sh
+    print_instruction "\nModify iptables"
+    sudo update-alternatives --set iptables /usr/sbin/iptables-legacy
+    sudo update-alternatives --set ip6tables /usr/sbin/ip6tables-legacy
+    sudo update-alternatives --set ebtables /usr/sbin/ebtables-legacy
 
-#sudo apt-mark hold kubelet kubeadm kubectl
+    print_instruction "\nCreating support file for k8s: kubernetes.list"
+    # Create a support file that will be copied to the nodes
+    echo "deb http://apt.kubernetes.io/ kubernetes-xenial main" | tee kubernetes.list
 
- # Do some cleanup
-print_instruction "\nDo some cleanup: autoremove\n"
-sudo apt-get -y autoremove
+    print_instruction "\nAdd Kubernetes repository to the RPi package lists"
+    sudo rm -f /etc/apt/sources.list.d/kubernetes.list
+    sudo cp -f kubernetes.list /etc/apt/sources.list.d/kubernetes.list
+    sudo echo 'Acquire::https::packages.cloud.google.com::Verify-Peer "false";' > apt.conf
+    sudo rm -f /etc/apt/apt.conf
+    sudo cp -f apt.conf /etc/apt/
 
+    print_instruction "\nUpdate & install kubelet kubeadm kubectl"
+    x=1
+    while [ $x -le 5 ]
+    do
+        print_instruction "Updating..."
+        count=sudo apt update | grep -c "404  Not Found"
+
+        if (( count >= 0 ))
+        then
+            x=10
+        fi
+
+        x=$(( $x + 1 ))
+
+    done
+    sudo apt install -y kubelet kubeadm kubectl
+
+    print_instruction "\nkubeadm init...\n"
+    sudo kubeadm init --ignore-preflight-errors=all --pod-network-cidr 10.244.0.0/16 --apiserver-advertise-address=$ip_addr_me
+
+    print_instruction "Configuring Kubernetes with local user"
+    echo $piid > piid.txt
+    sudo -u $piid sh ./_kube_config.sh
+
+    #sudo apt-mark hold kubelet kubeadm kubectl
+
+    # Do some cleanup
+    print_instruction "\nDo some cleanup: autoremove\n"
+    sudo apt-get -y autoremove
+
+    print_instruction "\n Done installing on the master node. \n"
+fi
+
+
+read -p "Press [Enter] key to setup worker nodes..."
 
 for ((i=0; i<$length; i++));
 do
@@ -91,11 +125,20 @@ do
     then
 
         # Run this code across all machines
-        print_instruction "\nRemoving swapfile on $host_target."
-        sudo sshpass -p $pword ssh $piid@$ip_target sudo dphys-swapfile swapoff
-        sudo sshpass -p $pword ssh $piid@$ip_target sudo dphys-swapfile uninstall
-        sudo sshpass -p $pword ssh $piid@$ip_target sudo update-rc.d dphys-swapfile remove
-        sudo sshpass -p $pword ssh $piid@$ip_target sudo apt-get -y purge dphys-swapfile
+        if [ $(swapon --show | grep -c "NAME") -gt 0 ]
+        then
+            print_instruction "\nRemoving swapfile on $host_target."
+            sudo sshpass -p $pword ssh $piid@$ip_target sudo dphys-swapfile swapoff
+            sudo sshpass -p $pword ssh $piid@$ip_target sudo dphys-swapfile uninstall
+            sudo sshpass -p $pword ssh $piid@$ip_target sudo update-rc.d dphys-swapfile remove
+            sudo sshpass -p $pword ssh $piid@$ip_target sudo apt-get -y purge dphys-swapfile
+        fi
+
+        if [ $(swapon --show | grep -c "NAME") -gt 0 ]
+        then
+            print_instruction "\nDiabling the swap file did not work...stopping."
+            exit 2
+        fi
 
         print_instruction "\nUpdate and install"
         sudo sshpass -p $pword ssh $piid@$ip_target sudo apt-get update
@@ -118,64 +161,13 @@ do
         sudo sshpass -p $pword ssh $piid@$ip_target sudo rm -f /etc/apt/sources.list.d/kubernetes.list
         sudo sshpass -p $pword ssh $piid@$ip_target sudo cp -f kubernetes.list /etc/apt/sources.list.d/kubernetes.list
 
+        sudo sshpass -p $pword ssh $piid@$ip_target sudo echo 'Acquire::https::packages.cloud.google.com::Verify-Peer "false";' > apt.conf
+        sudo sshpass -p $pword ssh $piid@$ip_target sudo rm -f /etc/apt/apt.conf
+        sudo sshpass -p $pword ssh $piid@$ip_target sudo cp -f apt.conf /etc/apt/
+
         print_instruction "Install kubelet kubeadm."
         sudo sshpass -p $pword ssh $piid@$ip_target sudo apt-get update
         sudo sshpass -p $pword ssh $piid@$ip_target sudo apt-get install -y kubelet kubeadm
-
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo rm -f /etc/apt/apt.conf
-        # sudo sshpass -p $pword scp apt.conf $piid@$ip_target:
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo cp -f apt.conf /etc/apt/
-
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo apt update
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo apt install -y kubelet kubeadm
-
-        # # print_instruction "Configuring Kubernetes with local user"
-        # # sudo sshpass -p $pword ssh $piid@$ip_target echo $piid > piid.txt
-        # # sudo sshpass -p $pword ssh $piid@$ip_target sudo -u $piid ./_kube_config.sh
-        # # sudo sshpass -p $pword ssh $piid@$ip_target rm piid.txt
-
-        # print_instruction "Locking: kubelet kubeadm" #kubectl"
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo apt-mark hold kubelet kubeadm #kubectl
-
-        # # Do some cleanup
-        # print_instruction "\nDo some cleanup: autoremove\n"
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo apt-get -y autoremove
-
-
-        # print_instruction "\nCopying kubernetes.list to worker machine."
-        # sudo sshpass -p $pword sudo scp "kubernetes.list" $piid@$ip_target:
-
-        # print_instruction "\nAdding link to Kubernetes repository and adding the APT key."
-        # # Add Kubernetes repository to the RPi package lists
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo rm -f /etc/apt/sources.list.d/kubernetes.list
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo cp -f kubernetes.list /etc/apt/sources.list.d/kubernetes.list
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo echo 'Acquire::https::packages.cloud.google.com::Verify-Peer "false";' > /etc/apt/apt.conf
-     # sudo sshpass -p $pword ssh $piid@$ip_target sudo apt-key adv --fetch-keys https://packages.cloud.google.com/apt/doc/apt-key.gpg
-
-        # print_instruction "\nUpdate and Upgrade"
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo apt update
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo apt -y upgrade
-
-
-        # print_instruction "Backing up /boot/cmdline.txt\n"
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo cp /boot/cmdline.txt /boot/cmdline_backup.txt
-
-        # sudo sshpass -p $pword ssh $piid@$ip_target echo "$(head -n1 /boot/cmdline.txt) cgroup_enable=cpuset cgroup_memory=1 cgroup_enable=memory" | sudo tee /boot/cmdline.txt
-
-        # print_instruction "\napt-get -y update\n"
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo apt-get -y update
-
-        # print_instruction "\napt-get -y install dnsutils kubeadm\n"
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo apt-get -y install dnsutils kubeadm
-
-        # # Do some cleanup
-        # print_instruction "\napt -y autoremove\n"
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo apt -y autoremove
-
-
-        # print_instruction "\nkubeadm init...\n"
-        # sudo sshpass -p $pword ssh $piid@$ip_target sudo kubeadm init --ignore-preflight-errors=all --pod-network-cidr 10.244.0.0/16 --apiserver-advertise-address=$ip_addr_me
-
 
     fi
